forge--read-topic: Require a match, as was already intended
When COLLECTION is a function, then REQUIRE-MATCH is ignored.

Because our collection function did not conform with the interface for
programmed completion functions, the behavior we got is what we would
expect if REQUIRE-MATCH were nil, even though it is t.

We can make our collection function conforming by passing it through
`completion-table-dynamic'.  (Even then REQUIRE-MATCH is still ignored,
but now nil is the values that is being ignored, which doesn't matter
here because we do require a match.)

Closes #637.