Improve kludge to support unsecured hosts
All that is necessary now is that users add the host to
`ghub-insecure-hosts'.  Defining a separate repository class is no
longer necessary.

Preserve support for the previous iteration of the kludge in
`forge--ghub-type-symbol'.  Not doing this would mean that users would
have to change the value of the class row of affected repositories in
the database, in addition to removing an entry from `forge-alist'.

Change the key that is looked up in `forge-alist' from GITHOST to
WEBHOST.  Since these formats are used for web and api urls, but not
git urls, it makes no sense to use the GITHOST as key.  The WEBHOST
and GITHOST are usually equal, but there is a small change that they
are not for some affected instances.

See #651 and #9.