forge--update-issue: Do not needlessly double down on id
In [1: 1220c0cce] we started setting the `id' slot using a dedicated
statement in `forge--update-issue', without doing the same in
`forge--update-pullreq'.  If it were necessary to do this, it would
be necessary in both cases.

Doing it is unnecessary because regarless of whether we retrieve an
existing object using `forge-get-{issue,pullreq}' or create a new object
using `forge-{issue,pullreq}', the `id' slot is already set, and cannot
change.

1: 2021-11-26 1220c0ccebc84e3689cd52df90dad6d08b307f10
   Support pulling existing topics in selective repositories again