forge--update-{issue,pullreq}: Always try to save people and labels
When a topic is fetched individually, i.e., outside a request that also
fetches data about the repository itself, then it can happen that the
topic data contains references to new users and labels, which do not
exist in the local database yet.  Because we only record the object id
for users and labels in topic tables as foreign keys, that can lead to
foreign key errors.

In the past we did not even attempt to update these values when fetching
individual topics.  That was an okay approach, because the expectation
was that users would always fetch everything.  Support for only fetching
individual topics was only added later, and was only intended to be used
in special cases.  This is changing (but still work in progress).

More importantly though, when we modify a topic, we now follow that up
by only fetching that topic from the forge again.  Previously we would
have fetched the complete repository, which was slow and unnecessary,
because at that time all we care about is making sure that one topic is
in sync.

In that case we must update all slots.  If the user modified, e.g., the
labels, then they expect that to take effect locally, without having to
explicitly fetch the complete repository.  We can even be sure that the
chosen labels are known locally already, else the user would not have
been able to select them.

When fetching an individual topic explicitly, it can still happen that
it references an unknown label or person, but that is the exception.
We now silently ignore these rare cases, which is better than to not
even attempting to store this data for the vast majority of cases for
which it would have worked.

Closes #654.