When first pulling notifications, mark none as pending
In [1: 715087aeb] we added new heuristics to interpret the "either done
or pending" status, which does take into account whether it is the first
time that all topics for a given repositories are being pulled.  We have
to use the same special case when notifications for a given forge are
being pulled for the first time.  In this case too, it is better to mark
all read topics as done, not pending.

1: 2024-03-18 715087aeb1ac35439015869004354471314dc211
   Implement new status update heuristics