forge--pull-topic: Don't fallback to refreshing another buffer
Calling `forge-refresh-buffer' with a dead buffer, is a no-op.
Take advantage of that.

(If we wanted to instead preserve the current behavior, we could
stop remembering the buffer and instead call `forge-refresh-buffer'
without any arguments, taking advantage of Ghub now making the
buffer current again, while calling the callback, which was current
when the request was made.)