Change bindings used to show the topic buffer and/or topic menu
Previously `forge-topic-menu' always showed not only the menu but also
the buffer.  Now it only shows the menu.  `forge-visit-this-topic'
continues to show only the buffer by default, but with a prefix argument
it now also shows the menu.

`forge-topics-menu' has become too crowded, because it features both
list filtering suffixes, as well as suffixes used to set topic slots.
Additionally, because `forge-topics-menu' and `forge-topic-menu' share
the latter set of suffixes, we had to give up on using single characters
for key bindings in both menus.

Subsequent commit will remove the setting suffixes from the list menu,
and shorten the used key sequences.

It should however remain possible to set topic slots directly from the
list buffer, so it has to be possible to bring up the topic menu from
the list buffer, without switching to the topic buffer.  That is what
this commit does.