forge-get-repository(demand): Use repository at point
Remove `forge-current-repository', which we previously for some commands
that should be able to act on the repository at point or else on the
"current" repository.  It turns out that most, if not all commands, that
act on "the appropriate" repository should be able to use the repository
at point, so do away with this distinction.

This instantly fixes many commands that previously did not work properly
in `forge-repository-list-mode' buffers.