forge-topics-menu: Pivot to forge-notifications-menu
We want "C-c C-c" to open the appropriate menu.  In Forge's own modes
we do this by remapping `forge--list-menu' to that menu.  We also want
"C-c C-c" to act that way in `magit-status-mode' (when the cursor is on
a topic or topic list), but here "C-c C-c" is bound to `magit-dispatch'
in the mode's keymap, so we have to use the section-specific keymap of
topic sections.

But that binding also takes effect for such sections in
`forge-notifications-mode' buffers, where we have to use
`forge-notifications-menu' instead (if we don't, we get an error).

Making `forge-topics-menu' to pivot to `forge-notifications-menu' when
invoked in `forge-notifications-mode' mode, is not elegant, but for now
it will have to do.