forge--topics-list-command: Remove misguided suffix class
All instances shared the same definition, which had to use
`transient-suffix-object' to get the suffix object.  Since that function
relies on the value of `this-command', this did not work when a command,
which uses this class, was invoked non-interactively.

That broke `forge-visit-this-repository', for example, which would have
had to set `this-command' to `forge-list-command' before calling that
function, which would be silly.