Make draft pull-requests italic
Add new face `forge-pullreq-draft' face, which makes draft pull-requests
italic by default.

Previously all done topics were italic, but that was an unnecessary
complication.  Pending topics are already bold (and unread topics are
bold and have a box around them), so it is enough for done topics to
just not be bold.

Closes #726.