forge--their-id: New function
Currently local object IDs derive from the upstream IDs.  We do not
use the upstream ID because those may not be unique across multiple
forges.  This might change in the future.

When using the Github REST API one can often use, e.g., a topic number
instead of an ID, but GraphQL mutations often require an ID, and we
are slowly migrating to that.

We could add new columns to many more tables to track "their ID", but
for now extract their ID from "our ID" if the `their-id' column/slot
does not already exist.

Yes, its a kludge.