forge--topic-parse-template-{,config}: Improve prompt format
We want to display both the "name" and the "about" value because
that's what is being shown in the web interface and so it might
be necessary to see both to know what to do.  But this also has
the potential to get very verbose, so we emphasize the "name".