forge-edit-post-hook: New hook
This hook is run in the buffer used to create a new discussion, issue
or pull-request, or to edit an existing topic or post, and replaces
`forge-create-pullreq-hook'.

The value of `buffer-file-name' can be used to determine what is being
created or edited.  For example, hook functions that should only do
something when creating a pull-request, could use:

  (when (string-prefix-p "new-pullreq"
                         (file-name-nondirectory buffer-file-name))
    ...)