Speed up fetching after replying to a topic or editing a post
After replying to a topic or editing an existing post, only fetch the
affected topic.  Doing a complete API pull (using `forge--pull') is not
only unnecessary in these cases, it also leads to a noticeable delay.

Continue to do a complete API pull after creating a topic, as that is
necessary to fetch the newly created topic (whose ID and number we
cannot know until the forge has assigned them).

Start potentially fetching just a single topic for Gitlab as well.
We didn't do that before because this isn't actually implemented for
Gitlab yet.  The `forge--pull-topic' method for Gitlab deals with that
limitation by pivoting to `forge--pull', so we don't have to deal with
that here as well.

When fetching a single pull-request using `forge--pull-topic', as we
do for "selective" repositories, fetch Git data, as we already did in
`forge--pull'.

Git data is fetched asynchronous, and when calling
`forge--maybe-git-fetch' without the optional BUFFER argument, as we do
in `forge--pull' and `forge--pull-topic', it does not cause any buffer
to be refreshed.  That is why we continue to do it after creating any
topic, even though it only really has to happen after creating a
pull-request.

Closes #786.