forge--{submit,cancel}-post-function: Break symmetry
These variables and the commands `forge-post-{submit,cancel}', that
use them, were added in the very first commit.  By now we know that
symmetry is not actually needed.

Remove `forge-cancel-post-function'.  We always want to use the same
canceling code -- which resides in `forge-post-cancel'.  (If a user
wants to change how the buffer is buried, they can set the buffer-
local value of `magit-bury-buffer-function', as for other Magit
buffers.

No longer use a dedicated error if `forge-submit-post-function' in
nil. That doesn't happen, and if it does it is a bug like any other,
not some expected issue that is likely to occur.