forge-get-topic(repo number-or-id): Actually support both number and id
Until [1: 243aa87d7] this method was broken for natural numbers,
afterwards for IDs.  Fix that by treating the two cases separately.

Also check the type of NUMBER in `forge-get-*(repo number)' methods,
so that we would get a more meaningful error going forward.  We have
to use `cl-assert' for this, because CLOS/EIEIO does not allow using
varying numbers of arguments for dispatch.  This is also why we have
to use a single method (repo number-or-id) instead of (repo number)
and (repo id) for `forge-get-topic', which is a bit sad.

Closes #791.

1: 2025-06-04 243aa87d700a815434bf31ab66a9e4cf1159ac50
   forge-get-topic: Pass along REPO argument